Apply an exception in the delete command

+++ HR.LeaveManagement.Application/Features/LeaveAllLocations/Handlers/Commands/DeleteLeaveAllLocationsCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using HR.LeaveManagement.Application.Features.LeaveAllLocations.Requests.Commands;
+﻿using HR.LeaveManagement.Application.Exceptions;
+using HR.LeaveManagement.Application.Features.LeaveAllLocations.Requests.Commands;
 using HR.LeaveManagement.Application.Persistence.Contracts;
 using MediatR;
 
@@ -16,6 +17,10 @@ namespace HR.LeaveManagement.Application.Features.LeaveAllLocations.Handlers.Com
         public async Task<Unit> Handle(DeleteLeaveAllLocationsCommand request, CancellationToken cancellationToken)
         {
             var leaveAllLocation = await _leaveAllLocationRepository.GetAsync(request.Id);
+
+            if (leaveAllLocation == null)
+                throw new NotFoundException(nameof(leaveAllLocation), request.Id);
+
             await _leaveAllLocationRepository.DeleteAsync(leaveAllLocation);
 
             return Unit.Value;
